src/dashboard/requirements.txt
textual
psycopg2-binary